requirements.txt
rich>=13.0
click>=8.0
yara-python>=4.3
pyyaml>=6.0
python-evtx>=0.7.4
xmltodict>=0.13
python-dateutil>=2.8
jinja2>=3.1
tqdm>=4.65
colorama>=0.4
tabulate>=0.9
pandas>=2.0
scipy>=1.11
pytest>=7.0
user-agents>=2.2
Whoosh>=2.7.4
drain3>=0.9.11
